+++ NEWS.md
@@ -3,6 +3,7 @@
 ## margins 0.2.24
 
 * The internal code of `cplot.lm()` has been refactored so that the actual plotting code now relies in non-exported utility functions, which can be used in other methods. This should make it easier to maintain existing methods and add new ones. (#49)
+* A new `cplot()` method for objects of class `"polr"` has been added (#49).
 
 ## margins 0.2.23